refactor: extract flow file operations to lib/flowOps.ts
Move flow create/duplicate/save/delete out of App.svelte into
src/lib/flowOps.ts, deduplicating the flow-name slug and the write-
register-open sequence shared by create and duplicate. Editor-local
flow UI state stays in App.svelte and is cleaned up by the thin
delete handler. No behavior change.
